added wrapper and default p font
