+++ src/App.tsx
@@ -3,14 +3,16 @@ import {Header} from './components/Header/Header';
 import {MainPage} from './pages/MainPage/MainPage';
 import { NotFoundPage } from './pages/NotFoundPage/NotFoundPage';
 import { PublishedVacancies } from './pages/PublishedVacancies/PublishedVacancies';
+import { MuiLoginForm } from './pages/MuiLoginForm/MuiLoginForm';
+import { MuiRegisterForm } from './pages/MuiRegisterForm/MuiRegisterForm';
 
 function App() {
   return (
     <>
       <Header />
       <Routes>
-        <Route path='/signin' element={<p>Signin</p>}/>
-        <Route path='/signup' element={<p>Signup</p>}/>
+        <Route path='/signin' element={<MuiLoginForm role='employer' />}/>
+        <Route path='/signup' element={<MuiRegisterForm role='employee' />}/>
         <Route path='/' element={<Navigate to='/employer/home'/>}/>
         <Route path='/employer/home' element={<PublishedVacancies />}/>
         <Route path='/employer/resumes' element={<MainPage />}/>
