+++ src/components/shared/AuthorizationWrapper/AuthorizationWrapper.module.scss
@@ -1,0 +1,9 @@
+.authWrapper {
+  width: 100%;
+  margin-top: 50px;
+  height: 80vh;
+  min-height: 400px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
