+++ src/components/shared/AuthorizationWrapper/AuthorizationWrapper.tsx
@@ -1,23 +1,14 @@
-// import { Stack } from '@mui/material';
-// import BasicSelect from '../MuiSelectButton/MuiSelectButton';
-// import styles from './AuthorizationWrapper.module.scss';
-// import { MuiRegisterForm } from '../../MuiRegisterForm/MuiRegisterForm';
-// import { Link } from 'react-router-dom';
-//
-// type TFormValues = {
-//   header: string
-// }
-// export function AuthorizationWrapper({ header, text, linkText }) {
-//   //Ю. Временный хардкод
-//   const role='employer'
-//   return(
-//     <div className={styles.authWrapper}>
-//     <Stack spacing={2} width={392}>
-//     <h1 className={styles.AuthorizationWrapper__header}>{header}</h1>
-//     <BasicSelect />
-//       <MuiRegisterForm role={role}/>
-//     </Stack>
-//       <p className={styles.test}>{text} <Link to='/{link}'>{linkText}</Link></p>
-//     </div>
-//   )
-// }
+import styles from './AuthorizationWrapper.module.scss';
+import { ReactNode } from 'react';
+
+interface IAuthorizationWrapper {
+  children: ReactNode;
+}
+export function AuthorizationWrapper({children}: IAuthorizationWrapper) {
+
+  return(
+    <div className={styles.authWrapper}>
+      {children}
+    </div>
+  );
+}
