+++ src/pages/MuiRegisterForm/MuiRegisterForm.tsx
@@ -1,14 +1,16 @@
-import { TextField, Button, Stack, Typography, Link} from '@mui/material';
+import { TextField, Button, Stack, Typography, Link } from '@mui/material';
 import { useForm } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
-import { schema } from './SchemaLogin';
-import { TFormValues, TLoginProps } from './TypesLogin';
-import BasicSelect from '../shared/MuiSelectButton/MuiSelectButton';
+import { TFormValues, TRegisterProps } from './TypesRegister';
+import { schema } from './SchemaRegister';
+import BasicSelect from '../../components/shared/MuiSelectButton/MuiSelectButton';
+import { AuthorizationWrapper } from '../../components/shared/AuthorizationWrapper/AuthorizationWrapper';
 
-export const MuiLoginForm: React.FC<TLoginProps> = ({role}) => {
+export const MuiRegisterForm: React.FC<TRegisterProps> = ({role}) => {
 
   const form = useForm<TFormValues>({
     defaultValues: {
+      name: '',
       email: '',
       password: '',
     },
@@ -21,12 +23,17 @@ export const MuiLoginForm: React.FC<TLoginProps> = ({role}) => {
     console.log(data);
   };
   return (
-    <div>
+    <AuthorizationWrapper>
       <Stack spacing={2} width={392}>
         <Typography>Авторизация</Typography>
         <BasicSelect />
       <form onSubmit={handleSubmit(onSubmit)} noValidate>
         <Stack spacing={2} width={392}>
+          <TextField placeholder='Фамилия Имя' type='text' {...register('name',
+            {required: 'Name is required'},
+          )}
+          error={!!errors.name}
+          helperText={errors.name?.message}/>
           <TextField placeholder='Почта' type='email' {...register('email',
             {required: 'Email is required'},
           )}
@@ -40,8 +47,8 @@ export const MuiLoginForm: React.FC<TLoginProps> = ({role}) => {
           <Button type='submit' variant='contained' color='primary' disabled={!isDirty || !isValid || (role === 'employee')}>Войти</Button>
         </Stack>
       </form>
-        <Typography>Новый пользователь? <Link href='#'>Зарегистрироваться</Link></Typography>
-        </Stack>
-    </div>
+        <Typography>Уже есть аккаунт? <Link href='/signin'>Войти</Link></Typography>
+      </Stack>
+    </AuthorizationWrapper>
   );
 };
