+++ src/pages/MuiRegisterForm/SchemaRegister.ts
