+++ src/pages/MuiRegisterForm/TypesRegister.ts
