+++ src/styles/base.scss
@@ -1,3 +1,5 @@
+@import './abstract/mixins';
+
 * {
   box-sizing: border-box;
 }
@@ -38,4 +40,9 @@ html {
   footer {
     flex: 0 0 auto;
   }
+
+  p {
+    @include font-text-reg;
+  }
+
 }
